Readme2 (#7)
* Added SkillCardComponent
A skill card contains a skill icon with the confidence level. A confidence level shows the confidence of the developer in that skill. It's value can be 1 to 4. For tools for which if it doesn't make sense to display confidence level, then setting confidence level to -1 simply removes it from the page for that particular skill.

Added a skills Page
This page display various technology and framework skills a developer has configured in the skills.json file. A skill is shown by the SkillCardComponent.

Added a readme Component
This component takes URL of the readme file and displays its content in an Angular Dialog.
- npm install --save ngx-markdown marked (for parsing readme to html)
- npm install --save prismjs (for showing code)
- npm instsall --save katex ( for parsing math content)


* Added text-nowrap to header links in the navigation bar

* Added Projects Page
Displays list of projects from projects.json file.